+++ crates/stackable-versioned-macros/src/codegen/container/enum.rs
@@ -3,7 +3,7 @@ use std::ops::Not;
 use darling::{util::IdentString, FromAttributes, Result};
 use proc_macro2::TokenStream;
 use quote::quote;
-use syn::ItemEnum;
+use syn::{Generics, ItemEnum};
 
 use crate::{
     attrs::container::NestedContainerAttributes,
@@ -46,6 +46,7 @@ impl Container {
         };
 
         Ok(Self::Enum(Enum {
+            generics: item_enum.generics,
             variants: versioned_variants,
             common,
         }))
@@ -82,6 +83,7 @@ impl Container {
         };
 
         Ok(Self::Enum(Enum {
+            generics: item_enum.generics,
             variants: versioned_variants,
             common,
         }))
@@ -96,12 +98,16 @@ pub(crate) struct Enum {
 
     /// Common container data which is shared between enums and structs.
     pub(crate) common: CommonContainerData,
+
+    /// Generic types of the enum
+    pub generics: Generics,
 }
 
 // Common token generation
 impl Enum {
     /// Generates code for the enum definition.
     pub(crate) fn generate_definition(&self, version: &VersionDefinition) -> TokenStream {
+        let (_, type_generics, where_clause) = self.generics.split_for_impl();
         let original_attributes = &self.common.original_attributes;
         let ident = &self.common.idents.original;
         let version_docs = &version.docs;
@@ -114,7 +120,7 @@ impl Enum {
         quote! {
             #(#[doc = #version_docs])*
             #(#original_attributes)*
-            pub enum #ident {
+            pub enum #ident #type_generics #where_clause {
                 #variants
             }
         }
@@ -133,6 +139,12 @@ impl Enum {
 
         match next_version {
             Some(next_version) => {
+                // TODO (@Techassi): Support generic types which have been removed in newer versions,
+                // but need to exist for older versions How do we represent that? Because the
+                // defined struct always represents the latest version. I guess we could generally
+                // advise against using generic types, but if you have to, avoid removing it in
+                // later versions.
+                let (impl_generics, type_generics, where_clause) = self.generics.split_for_impl();
                 let enum_ident = &self.common.idents.original;
                 let from_ident = &self.common.idents.from;
 
@@ -158,8 +170,10 @@ impl Enum {
                 Some(quote! {
                     #automatically_derived
                     #allow_attribute
-                    impl ::std::convert::From<#version_ident::#enum_ident> for #next_version_ident::#enum_ident {
-                        fn from(#from_ident: #version_ident::#enum_ident) -> Self {
+                    impl #impl_generics ::std::convert::From<#version_ident::#enum_ident #type_generics> for #next_version_ident::#enum_ident #type_generics
+                        #where_clause
+                    {
+                        fn from(#from_ident: #version_ident::#enum_ident #type_generics) -> Self {
                             match #from_ident {
                                 #variants
                             }