+++ crates/stackable-versioned/CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add basic support for generic types in struct and enum definitions ([#969]).
+
 ### Changed
 
 - BREAKING: Move `preserve_module` option into `options` to unify option interface ([#961]).
 
 [#961]: https://github.com/stackabletech/operator-rs/pull/961
+[#969]: https://github.com/stackabletech/operator-rs/pull/969
 
 ## [0.5.1] - 2025-02-14